Update bauerbot.mjs
fuck you
